requirements.txt
boto3==1.3.1
botocore==1.4.36
docutils==0.12
jmespath==0.9.0
python-dateutil==2.5.3
PyYAML==3.11
six==1.10.0
troposphere==1.7.0
